modify retry number type

+++ javaclient/src/main/java/me/laiyijie/job/android/io/swagger/model/Rule.java
@@ -17,7 +17,6 @@ import java.util.Objects;
 import com.google.gson.annotations.SerializedName;
 import io.swagger.annotations.ApiModel;
 import io.swagger.annotations.ApiModelProperty;
-import java.math.BigDecimal;
 import java.io.Serializable;
 
 /**
@@ -37,7 +36,7 @@ public class Rule implements Serializable {
   private String pattern = null;
 
   @SerializedName("retryTimes")
-  private BigDecimal retryTimes = null;
+  private Integer retryTimes = null;
 
   public Rule id(Integer id) {
     this.id = id;
@@ -93,7 +92,7 @@ public class Rule implements Serializable {
     this.pattern = pattern;
   }
 
-  public Rule retryTimes(BigDecimal retryTimes) {
+  public Rule retryTimes(Integer retryTimes) {
     this.retryTimes = retryTimes;
     return this;
   }
@@ -103,11 +102,11 @@ public class Rule implements Serializable {
    * @return retryTimes
   **/
   @ApiModelProperty(value = "重试次数")
-  public BigDecimal getRetryTimes() {
+  public Integer getRetryTimes() {
     return retryTimes;
   }
 
-  public void setRetryTimes(BigDecimal retryTimes) {
+  public void setRetryTimes(Integer retryTimes) {
     this.retryTimes = retryTimes;
   }
